reorganize js/diary folder
sorted components into sub-folders:
- StartEndLocations moved into js/diary/components
- LabelDetailsScreen moved into js/diary/details (it will be split up into smaller components)
- LabelListScreen moved into js/diary/list

- DiaryButton moved out of js/diary into /js/components

Now all that is at the root level is the main LabelTab component and 'helper' files
