+++ www/js/diary/cards/PlaceCard.tsx
@@ -16,7 +16,7 @@ import { getTheme } from "../../appTheme";
 import { DiaryCard, cardStyles } from "./DiaryCard";
 import { useAddressNames } from "../addressNamesHelper";
 import useDerivedProperties from "../useDerivedProperties";
-import StartEndLocations from "../StartEndLocations";
+import StartEndLocations from "../components/StartEndLocations";
 
 type Props = { place: {[key: string]: any} };
 const PlaceCard = ({ place }: Props) => {
