+++ www/js/diary/cards/UntrackedTimeCard.tsx
@@ -15,7 +15,7 @@ import { useTranslation } from "react-i18next";
 import { DiaryCard, cardStyles } from "./DiaryCard";
 import { useAddressNames } from "../addressNamesHelper";
 import useDerivedProperties from "../useDerivedProperties";
-import StartEndLocations from "../StartEndLocations";
+import StartEndLocations from "../components/StartEndLocations";
 
 type Props = { triplike: {[key: string]: any}};
 const UntrackedTimeCard = ({ triplike }: Props) => {
