+++ www/js/survey/enketo/AddNoteButton.tsx
@@ -8,7 +8,7 @@
 */
 
 import React, { useEffect, useState, useContext } from "react";
-import DiaryButton from "../../diary/DiaryButton";
+import DiaryButton from "../../components/DiaryButton";
 import { useTranslation } from "react-i18next";
 import moment from "moment";
 import { LabelTabContext } from "../../diary/LabelTab";
